feat(github): default branch resolves to lowest-tier environment branch
The settings baseline now sets the repository default branch to the
lowest-tier environment branch that exists (dev > staging > main);
repos with none of them keep their current default. Overridable per
repo via .lisa.config.json github.settings.default_branch.

Settings-script tests move to their own file to stay under max-lines.
